Revert "Evan/oauth"
