+++ frontend/src/app/(ui)/dashboard/layout.tsx
@@ -1,4 +1,3 @@
-
 import React from "react";
 
 import ResponsiveAppBar from "../app-bar";
