+++ frontend/src/app/(ui)/dashboard/page.tsx
@@ -1,4 +1,3 @@
-
 import { Box, Container, Paper, Typography } from '@mui/material';
 
 export default function DashboardPage() {
