--- frontend/src/app/(ui)/dashboard/user-menu.tsx
@@ -1,56 +1,0 @@
-'use client';
-
-import { authConfig } from '@/app/api/auth/[...nextauth]';
-import { useSession } from 'next-auth/react';
-
-import * as React from 'react';
-import { Box, Menu, MenuItem, Typography, Tooltip, IconButton, Avatar } from '@mui/material';
-import {test} from '@/app/(lib)/actions';
-
-const settings = [ 'Profile', 'Log Out' ];
-
-export default function UserMenu() {
-  const [anchorElUser, setAnchorElUser] = React.useState<null | HTMLElement>(null);
-
-  const handleOpenUserMenu = (event: React.MouseEvent<HTMLElement>) => {
-    setAnchorElUser(event.currentTarget);
-  };
-
-  const handleCloseUserMenu = () => {
-    setAnchorElUser(null);
-  };
-
-  const { data: session, status } = useSession();
-
-  return (
-    <Box sx={{ flexGrow: 0 }}>
-    <Tooltip title="Open settings">
-      <IconButton onClick={handleOpenUserMenu} sx={{ p: 0 }}>
-        {session && <Avatar src={session?.user?.image?.toString()} alt="" />}
-      </IconButton>
-    </Tooltip>
-    <Menu
-      sx={{ mt: '45px' }}
-      id="menu-appbar"
-      anchorEl={anchorElUser}
-      anchorOrigin={{
-        vertical: 'top',
-        horizontal: 'right',
-      }}
-      keepMounted
-      transformOrigin={{
-        vertical: 'top',
-        horizontal: 'right',
-      }}
-      open={Boolean(anchorElUser)}
-      onClose={handleCloseUserMenu}
-    >
-      {settings.map((setting) => (
-        <MenuItem key={setting} onClick={handleCloseUserMenu}>
-          <Typography textAlign="center">{setting}</Typography>
-        </MenuItem>
-      ))}
-    </Menu>
-  </Box>
-  );
-}
