--- frontend/src/app/AuthProvider.tsx
@@ -1,14 +1,0 @@
-'use client';
-
-import { SessionProvider } from 'next-auth/react';
-import React from 'react';
-
-export default function AuthProvider({ children }: {
-  children: React.ReactNode
-}) {
-  return (
-    <SessionProvider>
-      {children}
-    </SessionProvider>
-  );
-}
