--- frontend/src/app/api/auth/[...nextauth].ts
@@ -1,45 +1,0 @@
-import { NextAuthOptions, User, getServerSession } from 'next-auth';
-
-// import CredentialsProvider from 'next-auth/providers/credentials';
-import GoogleProvider from 'next-auth/providers/google';
-import { useSession } from 'next-auth/react';
-
-import { redirect, useRouter } from 'next/navigation';
-
-export const authConfig: NextAuthOptions = {
-  providers: [
-    /*CredentialsProvider({
-      name: "Sign In",
-      credentials: {
-        email: {
-          label: "Email",
-          type: "email",
-          placeholder: "example@example.com",
-        },
-        password: { label: "Password", type: "password" },
-      },
-
-      async authorize(credentials) {
-        if (!credentials || !credentials.email || !credentials.password)
-          return null;
-
-        const dbUser = await {}.user.findFirst({
-          where: { email: credentials.email },
-        });
-        
-        // USE BCRYPT
-        if (dbUser && dbUser.password === credentials.password) {
-          const { password, createdAt, id, ...dbUserWithoutPassword } = dbUser;
-          return dbUserWithoutPassword as User;
-        }
-
-        return null;
-      },
-    }),*/
-
-    GoogleProvider({
-      clientId: process.env.GOOGLE_CLIENT_ID as string,
-      clientSecret: process.env.GOOGLE_CLIENT_SECRET as string,
-    }),
-  ],
-};
