--- frontend/src/app/api/auth/[...nextauth]/route.ts
@@ -1,6 +1,0 @@
-import { authConfig } from '@/app/api/auth/[...nextauth]';
-import NextAuth from 'next-auth/next';
-
-const handler = NextAuth(authConfig);
-
-export { handler as GET, handler as POST };
