--- frontend/src/middleware.js
@@ -1,3 +1,0 @@
-export { default } from "next-auth/middleware";
-
-export const config = { matcher: ['/dashboard/:path*']}
